refactor(json): bring keel-api-jackson into keel-core
