+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/jackson/PropertyNamePolymorphicDeserializer.kt
@@ -1,4 +1,4 @@
-package com.netflix.spinnaker.keel.json
+package com.netflix.spinnaker.keel.jackson
 
 import com.fasterxml.jackson.databind.DeserializationContext
 import com.fasterxml.jackson.databind.JsonNode
