+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/jackson/ResourceKindDeserializer.kt
@@ -1,4 +1,4 @@
-package com.netflix.spinnaker.keel.json
+package com.netflix.spinnaker.keel.jackson
 
 import com.fasterxml.jackson.core.JsonParser
 import com.fasterxml.jackson.databind.DeserializationContext
