+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/jackson/mixins/DeliveryArtifactMixin.kt
@@ -1,4 +1,4 @@
-package com.netflix.spinnaker.keel.json.mixins
+package com.netflix.spinnaker.keel.jackson.mixins
 
 import com.fasterxml.jackson.annotation.JsonProperty
 import com.fasterxml.jackson.annotation.JsonProperty.Access.WRITE_ONLY
