+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/jackson/mixins/ResourceKindMixin.kt
@@ -1,9 +1,9 @@
-package com.netflix.spinnaker.keel.json.mixins
+package com.netflix.spinnaker.keel.jackson.mixins
 
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize
 import com.fasterxml.jackson.databind.annotation.JsonSerialize
 import com.fasterxml.jackson.databind.ser.std.ToStringSerializer
-import com.netflix.spinnaker.keel.json.ResourceKindDeserializer
+import com.netflix.spinnaker.keel.jackson.ResourceKindDeserializer
 
 @JsonSerialize(using = ToStringSerializer::class)
 @JsonDeserialize(using = ResourceKindDeserializer::class)
