+++ keel-core/src/main/kotlin/com/netflix/spinnaker/keel/jackson/mixins/SubnetAwareRegionSpecMixin.kt
@@ -1,4 +1,4 @@
-package com.netflix.spinnaker.keel.json.mixins
+package com.netflix.spinnaker.keel.jackson.mixins
 
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonInclude.Include.NON_EMPTY
