+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/ec2/jackson/ImageProviderDeserializer.kt
@@ -21,7 +21,7 @@ import com.netflix.spinnaker.keel.api.ec2.ArtifactImageProvider
 import com.netflix.spinnaker.keel.api.ec2.ImageProvider
 import com.netflix.spinnaker.keel.api.ec2.JenkinsImageProvider
 import com.netflix.spinnaker.keel.api.ec2.ReferenceArtifactImageProvider
-import com.netflix.spinnaker.keel.json.PropertyNamePolymorphicDeserializer
+import com.netflix.spinnaker.keel.jackson.PropertyNamePolymorphicDeserializer
 import com.netflix.spinnaker.kork.exceptions.UserException
 
 internal class ImageProviderDeserializer :
