+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/ec2/jackson/SecurityGroupRuleDeserializer.kt
@@ -4,7 +4,7 @@ import com.netflix.spinnaker.keel.api.ec2.CidrRule
 import com.netflix.spinnaker.keel.api.ec2.CrossAccountReferenceRule
 import com.netflix.spinnaker.keel.api.ec2.ReferenceRule
 import com.netflix.spinnaker.keel.api.ec2.SecurityGroupRule
-import com.netflix.spinnaker.keel.json.PropertyNamePolymorphicDeserializer
+import com.netflix.spinnaker.keel.jackson.PropertyNamePolymorphicDeserializer
 
 internal class SecurityGroupRuleDeserializer :
   PropertyNamePolymorphicDeserializer<SecurityGroupRule>(SecurityGroupRule::class.java) {
